backend/requirements.txt
# ./backend/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
sqlalchemy==1.4.27
psycopg2-binary==2.9.2
boto3==1.20.20
azure-identity==1.7.1
azure-mgmt-resource==21.0.0
azure-mgmt-compute==26.0.0
azure-mgmt-storage==19.0.0
google-cloud-resource-manager==1.3.3
google-cloud-compute==1.0.0
google-cloud-storage==1.43.0
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
passlib==1.7.4
bcrypt==3.2.0
python-dotenv==0.19.2